LPD-101253 Share the Top Assets card in shared components
